Refactor container `ps` command flagging process
